+++ console2/src/components/molecules/AnsibleHostList/index.tsx
@@ -19,14 +19,15 @@
  */
 
 import * as React from 'react';
-import { Header, Input, Menu, Modal } from 'semantic-ui-react';
+import { Input, Menu, Modal, Table } from 'semantic-ui-react';
 
 import { AnsibleEvent, AnsibleStatus, ProcessEventEntry } from '../../../api/process/event';
-import { AnsibleTaskList } from '../index';
+import { AnsibleTaskList, HumanizedDuration } from '../index';
 
 export interface AnsibleHostListEntry {
     host: string;
     status: AnsibleStatus;
+    duration: number;
 }
 
 interface State {
@@ -39,11 +40,14 @@ interface Props {
     selectedStatus?: AnsibleStatus;
 }
 
+const compareByHost = (a: AnsibleHostListEntry, b: AnsibleHostListEntry) =>
+    a.host > b.host ? 1 : a.host < b.host ? -1 : 0;
+
 const applyFilter = (
     hosts: AnsibleHostListEntry[],
     selectedStatus?: AnsibleStatus,
     filter?: string
-): string[] => {
+): AnsibleHostListEntry[] => {
     let result = [...hosts];
 
     if (selectedStatus) {
@@ -55,7 +59,7 @@ const applyFilter = (
         result = result.filter(({ host }) => host.toLowerCase().indexOf(f) >= 0);
     }
 
-    return Array.from(new Set(result.map((e) => e.host))).sort();
+    return result.sort(compareByHost);
 };
 
 class AnsibleHostList extends React.Component<Props, State> {
@@ -64,7 +68,7 @@ class AnsibleHostList extends React.Component<Props, State> {
         this.state = {};
     }
 
-    renderHostItem(host: string, idx: number) {
+    renderHostItem(host: string, duration: number, idx: number) {
         const { hostEventsFn } = this.props;
         return (
             <Modal
@@ -73,9 +77,12 @@ class AnsibleHostList extends React.Component<Props, State> {
                 size="fullscreen"
                 dimmer="inverted"
                 trigger={
-                    <Menu.Item key={idx} name={host}>
-                        {host}
-                    </Menu.Item>
+                    <Table.Row>
+                        <Table.Cell>{host}</Table.Cell>
+                        <Table.Cell>
+                            <HumanizedDuration value={duration} />
+                        </Table.Cell>
+                    </Table.Row>
                 }>
                 <Modal.Content>
                     <AnsibleTaskList title={host} events={hostEventsFn(host)} />
@@ -90,25 +97,34 @@ class AnsibleHostList extends React.Component<Props, State> {
 
         return (
             <>
-                <Menu attached="top" borderless={true} secondary={true}>
-                    <Menu.Item>
-                        <Header as="h4">Hosts by Status {selectedStatus}</Header>
-                    </Menu.Item>
-                    <Menu.Item position="right">
+                <Menu attached="top" borderless={true} tabular={true}>
+                    <Menu.Item position={'right'}>
                         <Input
                             type="text"
                             icon="filter"
                             size="small"
-                            placeholder="Search"
+                            placeholder="Host"
                             onChange={(e, { value }) => this.setState({ filter: value })}
                         />
                     </Menu.Item>
                 </Menu>
-                <Menu attached="bottom" vertical={true} fluid={true}>
-                    {applyFilter(hosts, selectedStatus, filter).map((host, idx) =>
-                        this.renderHostItem(host, idx)
-                    )}
-                </Menu>
+
+                <Table celled={true} attached="bottom" selectable={true}>
+                    <Table.Header>
+                        <Table.Row>
+                            <Table.HeaderCell singleLine={true}>
+                                Hosts by Status {selectedStatus}
+                            </Table.HeaderCell>
+                            <Table.HeaderCell singleLine={true}>Duration</Table.HeaderCell>
+                        </Table.Row>
+                    </Table.Header>
+
+                    <Table.Body>
+                        {applyFilter(hosts, selectedStatus, filter).map((host, idx) =>
+                            this.renderHostItem(host.host, host.duration, idx)
+                        )}
+                    </Table.Body>
+                </Table>
             </>
         );
     }