+++ console2/src/components/molecules/index.tsx
@@ -28,6 +28,7 @@ export { default as DonutChart } from './DonutChart';
 export { default as EntityRenameForm } from './EntityRenameForm';
 export { default as GlobalNavMenu, GlobalNavTab } from './GlobalNavMenu';
 export { default as Highlighter } from './Highlighter';
+export { default as HumanizedDuration } from './HumanizedDuration';
 export { default as LocalTimestamp } from './LocalTimestamp';
 export { default as NewAPITokenForm, NewAPITokenFormValues } from './NewAPITokenForm';
 export { default as NewProjectForm, NewProjectFormValues } from './NewProjectForm';