+++ plugins/tasks/ansible/src/main/resources/com/walmartlabs/concord/plugins/ansible/strategy/concord_free.py
@@ -24,5 +24,7 @@ class StrategyModule(StrategyModule_free):
             super(StrategyModule, self)._queue_task(host, _task, task_vars, play_context)
             return
 
+        self._tqm.send_callback('concord_on_task_start', host, task)
+
         super(StrategyModule, self)._queue_task(host, task, task_vars, play_context)