+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-console: display flow and ansible event duration;
+- ansible: record "pre" and "post" task events separately;
 - concord-console: new API key management UI;
 - concord-server: optional API key expiration;
 - concord-server, concord-console: new method to validate a repo